feat(tcp): refactor handshake into separate file
